python-gatekeeper/requirements.txt
Flask==1.0.2
honeycomb-beeline==1.0.0
python-dateutil==2.7.3